requirements.txt
imageio
Pillow
wandb
tqdm
rich
fire

